requirements.txt
requests>=2.20.0
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
elasticsearch>=5.0.0,<6.0.0
future==0.17.1
idna==2.6
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
splunk-sdk
urllib3==1.22
yara-python
